requirements.txt
psycopg2-binary  
confluent-kafka
simplejson
pyspark==3.5.7
streamlit
streamlit-autorefresh
kafka-python
matplotlib
